backend/requirements.txt
# Flask & Extensions
Flask==3.1.2
Flask-CORS==5.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
gunicorn==23.0.0

# Database
SQLAlchemy==2.0.44
alembic==1.17.2
psycopg2-binary==2.9.11

# Security
PyJWT==2.10.1
bcrypt==4.2.1

# Development
pytest==8.3.5

